backend/wikipedia-data-ingestion/requirements.txt
# Core dependencies
azure-functions>=1.17.0
azure-identity>=1.12.0
azure-storage-blob>=12.16.0
azure-search-documents>=11.4.0
openai>=1.0.0

# Data processing
datasets>=2.13.0
transformers>=4.30.0
nltk>=3.8.1
regex>=2022.10.31

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Monitoring and logging
opencensus-ext-azure>=1.1.9
applicationinsights==0.11.10

# Utilities
python-dotenv>=1.0.0
